update JDK 17

+++ src/main/java/com/bpi/model/dto/CoindeskBpiDTO.java
@@ -7,6 +7,8 @@ import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
+import java.math.BigDecimal;
+
 /**
  * for Coindesk 用的 rate_float
  * 
@@ -35,6 +37,6 @@ public class CoindeskBpiDTO {
 	String rate;
 	
 	@JsonProperty("rate_float")
-	Double rateFloat;
+	BigDecimal rateFloat;
 
 }
